split up dateslider components
